Unit tests restructure
